+++ engine/c/kvdk_batch.cpp
@@ -1,3 +1,7 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ * Copyright(c) 2021-2022 Intel Corporation
+ */
+
 #include "kvdk_c.hpp"
 
 extern "C" {
