+++ engine/c/kvdk_hash.cpp
@@ -1,3 +1,7 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ * Copyright(c) 2021-2022 Intel Corporation
+ */
+
 #include <cassert>
 #include <cstring>
